DO NOT SUBMIT - Zoekt Polymer Frontend
This change should not be submitted. It shall be used to judge
if we want to continue working on a Polymer frontend for Zoekt
or if we stick to the current plain HTML implementation.

If we continue with Polymer, the Polymer App in this change needs
more work as parts of it are just hacked together to create a
working prototype.

The API cmd/zoekt-webserver/main.go was never meant to be published
or go in prod. It is just a hack to be able to show a working
prototype. If we continue with Polymer, a proper API needs to be
implemented and the whole topic needs more thought.

Change-Id: Ie1f5249838a140e6d837f63f13d75a9e98aac63e