Update patch set 6
Patch Set 6:

so, I don't think we want to  merge this as-is; it looks prettier, but it adds a bunch of overhead to load the pages. 

I would be very interested in working on a search-as-you type experience; maybe we can sit down some day to make that happen.

Patch-set: 6