Update patch set 1
Patch Set 1: Code-Review-1

(4 comments)

I think we should change the way logs are handled here. I don't like that fact, that we have to restart the webserver.

An easy fix would be to switch to an hourly log file schema: logs.2016060912 for the 12th our of today. The maxLogAge is then a multiple of 1 hour.

Another fix would be to work with a file mutex that is acquired before a log message is written (logs can be stored in an in-memory data structure first and written to disk every now and then). Another process can clean the same log file by acquiring the mutex first.

Patch-set: 1
Label: Code-Review=-1