Update patch set 1
Patch Set 1:

(3 comments)

for logs, it's best to avoid buffering, because if the program goes down, it takes the buffers into the grave.  This is the advantage of the current approach: since the OS does the stderr redirection, the log always captures everything. I wanted to avoid changing the webserver (letting it continue to log to stderr), but maybe that is silly.

Patch-set: 1
Label: Verified=0