Handle case folding by generating ngram variants.
When given a query "Abc" and corpus "Xyz ABC", before, we would
transform the corpus and search query into "xyz abc" and "abc"
respectively. This works for ASCII because there upper and lower case
have a 1-1 mapping, it is problematic for UTF-8, where upper and
lowercase runes may have different widths.

Instead, when searching with case folding for "Abc", we look for {"Abc",
"abc", "ABc", "aBc", ...} in the original text.

Beyond opening the path for unicode case handing, this has several other
advantages:

* we don't have to generate and read case bits.

* we don't have to regenerate the original from its lowercased version
  for snippets, which was a bottleneck.

* we have to consider fewer documents in case of case-sensitive
  searches, since we can look for ngrams with the right casing.

Change-Id: I54b575d347819e7a13f9cbb8095e9bca40de2f37