Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> We should fix getting project name dropped from the name, but why do you care about how long the filenames are?

Nothing important. Just general readability (in logs etc). Also, depending on the naming, it is not impossible to run into a FS limit on path-names:

  /this/is/where/we/store/gits/gittools.se.axis.com/gerrit/gitweb?p=packages/param/network/interface/another_namespace_parameter/this-is-a-really-long-git-name-since-it-comes-frome-a-product-name_v14.00000.zoekt

While I changed some components of the path for various reasons, we have gits that will generate filenames of this length. That's 210 characters. A bit too close to the 255 chars limit that quite a few FS have.

Patch-set: 1