Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> we can truncate the name to some limit (200 chars?) and append a SHA1 for disambiguation.

That sounds better. At least we wont see filename collisions.

Though I still think that just using the repo name is a better way.

For instance, in a few weeks we will roll out gitiles, which we want to use as the default zoekt.web-url. Hence we will see index duplicates, even though the actual git we index is unchanged.

Patch-set: 1