Limit shard filenames to 200 bytes.
This avoids problems when using it on filesystems with limited file
names, eg. NFS.

For performance reasons, searches should not be served from NFS, but a
network filesystem is useful for storing the master set of index files
when the search server is sharded.

Bump FormatVersion to avoid loading shards from their old and new
filenames at the same time.

Change-Id: I2e712ed764526acdb6c5ce7733023b92444d530a