Update patch set 3
Patch Set 3:

Sorry for the previous comment that ended with a cliff hanger. I stumbled upon a new issue when trying out this patch on our Zoekt-server and started to write a comment about it. But since it was not related to this change, I decided to removed it (but apparently accidentally missed a few chars). If curious, this was the issue: https://github.com/google/zoekt/issues/29

Patch-set: 3
Reviewer: Gustaf Lundh <1021592@173816e5-2b9a-37c3-8a2e-48639d4f1153>