Update patch set 2
Patch Set 2:

For gerrit (which was the origin of this idea), there is an API check.access

  https://gerrit-review.googlesource.com/Documentation/rest-api-projects.html#check-access

the rough idea was that each shard would contact gerrit to verify readability for (repo, ref, user) tuples, which gets cached for a while. This may cause a lot of load on Gerrit, but there is probably no better way to do it for the Gerrit case, as its permission model is really complicated and must be evaluated for each user separately.

If you have some sort of system that can provide a full list of repos that are visible to a user, then a special query construct that can be handled specially in shardedSearcher indeed is better. 

That said, the REST API has not been thought through very well, and it duplicates the API surface, so if you're really building something, it would be good to double check that the REST package is the way to go.

Patch-set: 2