Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 1:
> > 
> > Question: the line number anchors on the /print page now put the line under the search box. Is there a way to keep the current layout, but convince the browser that the top of the scrollable content starts below the search box?
> 
> It sounds like you are trying to do similar to what I did here: https://gerrit-review.googlesource.com/c/zoekt/+/137213

(looking)

aha! I understand what is going on (which is quite the achievement given my knowledge of HTML :)

However, I also want to highlight the selected line through 

  :target {
     background: .. 
  }

and that doesn't work with your trick since the trick makes the actual anchor invisible. I tried looking for other ways to make the anchor be in the center of the page, but it seems that is impossible. 

Can you recommend a way to make the content something separate (so it doesnt go under the navbar?)

Patch-set: 2
Reviewer: hanwen <1024147@173816e5-2b9a-37c3-8a2e-48639d4f1153>