Update patch set 1
Patch Set 1:

> can you send the other change too (you can just commit on top of this commit)?

I haven't started on that, was about too. It would be good to get some guidance before investing time in it though. Two options to exclude:

1. Exclude long lines from the indexfile (better?)
2. Filter out long lines in results.


> I was going to suggest that it should be a field in IndexBuilderOption , but we don't have that struct (yet).

That should be easy to add if you would rather it was configurable. But I think this is a poor heuristic for excluding files due to the false positives.

Patch-set: 1
Reviewer: Gerrit User 1060017 <1060017@173816e5-2b9a-37c3-8a2e-48639d4f1153>