query: Handle exclusion ranges in LowerRegexp
An exclusion class is something like [^A-Z]. We need to specially handle it
since the user intention of [^A-Z] should map to [^a-z]. If we use the
normal mapping logic, we will do nothing since [a-z] is in [^A-Z].

Unfortunetly this resulted in some pretty hairy logic. This problem maps to
manipulating line intersections. There is possibly a way to write this in a
clearer way, but for now we have many more unit tests.

Change-Id: I157342417fa79cdb39645302718d4ae2c9a5aabf