Include files without text on index
Instead of excluding files without text, those files are included, changing the language to 'binary',
allowing the match search of strings or chars present in those kind of files, but without print the line content.
For example, looking for '\x00 r:linux' on beagleboard/linux project, it is possible localise pdf files,
hiding the results on web and printing the '<probably not text>' in the zoekt results.

Resolves: #57

Change-Id: I1aa08ff45c702005a2f23c9e5c590023cb5fac2f