Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> I added logic to store the large files list as part of index metadata so we can re-index if the list changed (when -incremental=true).


I appreciate the effort, but I think it should go into a different change. If we do this, we should also support the size threshold, and other flags that might influence what gets indexed (eg. whether to use CTags). That in itself suggests that we should create a separate cli/ library to create flags across all the indexers and then centrally handle this. All of this should be in a separate change.

I was thinking of Keegan's idea of deriving regex from glob, but I think we can change our minds about this in the future.

Patch-set: 6
Reviewer: Gerrit User 1024147 <1024147@173816e5-2b9a-37c3-8a2e-48639d4f1153>