Update patch set 7
Patch Set 7:

> I appreciate the effort, but I think it should go into a different change. If we do this, we should also support the size threshold, and other flags that might influence what gets indexed (eg. whether to use CTags). That in itself suggests that we should create a separate cli/ library to create flags across all the indexers and then centrally handle this. All of this should be in a separate change.
> 
> I was thinking of Keegan's idea of deriving regex from glob, but I think we can change our minds about this in the future.

Makes sense! I've reverted back to the equivalent of patch set 5, removing all that logic. A separate package for this sounds great. I'd be happy to work on that change as well as we'd like the support for that.

Patch-set: 7
Reviewer: Gerrit User 1081869 <1081869@173816e5-2b9a-37c3-8a2e-48639d4f1153>