ctags: kill subprocess on scanner errors
github.com/sirthias/borer contains a JSON file with 100k '[' chars,
This caused a "token too long" error from bufio, which left the
subprocess alive. Before, we waited for the subprocess to die, which
never happened.

This causes the indexing to fail for this repo, but that is better
than stopping the complete indexing pipeline.

Change-Id: I4ee370ae5e531fc8fde0256eac3d280771594c73