Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> why?

Personally I find the code sometimes a bit hard to understand due to the boolean that is passed around for filename vs content. But we (Sourcegraph) are experimenting with using Zoekt to index more ctags information. This change is probably a prerequisite for what we do to add that support.

On that note, this may be something upstream zoekt is interested in. Are ya?

Patch-set: 1