Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > why?
> 
> In the future, there might be a need to search in other corpuses aside from file names and file content. This change would make the code more flexible to allow for those additions. In particular, I've been thinking about Zoekt supporting regular expression + <2 length queries for symbol search. In this case, we could easily add `ScopeSymbol` to the enum and implement it from there (add additional postings list for symbols).

that makes some sense, but wouldn't it be better to start with the hard part of that project, ie. implementing symbol search? 

One of the things I don't like about this change is that it breaks the existing API, and makes it more cumbersome to use (the default value stops doing something useful). I realize that SG might be the only user of the API, but it is still something I would like to avoid.

Patch-set: 1
Reviewer: Gerrit User 1028062 <1028062@173816e5-2b9a-37c3-8a2e-48639d4f1153>