Update patch set 1
Patch Set 1:

(2 comments)

Found an issue when testing it out. The matches returned are always from the beginning of the symbol. For example on the `ripgrep` repo I searched `sym:mat.h case:yes`. The results for the symbol `only_matching` highlighted `only_`. Can you add a test case for this?

Patch-set: 1
CC: Gerrit User 1082084 <1082084@173816e5-2b9a-37c3-8a2e-48639d4f1153>