Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I'm sorry to see this is abandoned. Were you waiting for me?

Sorry, I intended to submit a new Gerrit change containing the newer changes with the alternative approach with treating symbols as docs in an iterator, but it's probably better to update this one instead. This is definitely not the most performant approach (from the overhead of so many regex match calls) which is why I initially abandoned it. I'll include benchmarks in the change later this week.

Patch-set: 3
Reviewer: Gerrit User 1085159 <1085159@173816e5-2b9a-37c3-8a2e-48639d4f1153>