all: move incremental indexing decision into build options
Currently build.Options returns a list of indexed versions. This is used in
each command to decide if we should incrementally index. However, it has been
updated over time to also do extra checks for incremental indexing. If one of
those checks fail, it returns nil to force the indexing.

This commit updates the function to encapsulate what we want, the decision of
should we index given the build options.

Change-Id: I3a8b9073408cb785a299bc287af1e10570415c96