Update patch set 5
Patch Set 5:

(4 comments)

Sorry, I've been having trouble with the Gerrit UI. I'm not very familiar with this tool. I updated the code with your suggestion. Sorry it took a while to respond back too. I tried leaving a comment a couple weeks ago but it was saved as a draft by mistake.

The checks seem to fail but the commit message hasn't changed significantly. I added the commit-msg git hook for generating Change IDs so I'm not sure what could be going wrong. Are the checks indicating actual errors?

Patch-set: 5