build: fsync shards and indexdir
This ensures data is synced to disk, as well as any renames and removals
we do in the process of updating indexes. We now call fsync once per
shard (in its own goroutines) and once finally for the index
directory. This shouldn't impact performance, since time should be
dominated by the writing of the large shards.

Change-Id: I815f2e103f01582222c1c07d1fb3feeb78d9e0cc