improve performance for content searches
For content searches trying to match multiple terms on the same line, we
check whether the matches of the individual terms intersect before
calling the regex engine. If they don't intersect, we skip the document.

This optimisation is useful whenever terms of the query appear often in
the same document but rarely on the same line.

Local tests (master vs. this change) on a set of medium-sized repos
showed that for queries such as `(func)(?-s:.)*?(err)` latencies reduce
by around 30%. For queries where the terms rarely appear on the same
line, such as `(field)(?-s:.)*?(bar)`, even up to 86%.`

Change-Id: Ic4e4c3e7a4ad61ec56efada35a70aa14d0766ff7